More refactoring, comments and new README.md
