+++ README.md
@@ -3,19 +3,136 @@
 [![Coverage Status](https://coveralls.io/repos/github/d4rken/ommvplib/badge.svg)](https://coveralls.io/github/d4rken/ommvplib)
 [![Build Status](https://travis-ci.org/d4rken/ommvplib.svg?branch=master)](https://travis-ci.org/d4rken/ommvplib)
 
-This library offers a structure for building Android apps follow MVP principles. Adhering to this structure yields a modular and well testable app.
+This library offers a structure that helps build Android apps that follow MVP principles.
 
-The core concept consists of Dagger2's `AndroidInjector` while using Android's `Loader` to retain presenters and components across lifecycle events
+The core concept utilizes a `Loader` to store the `Presenter` during configuration (e.g. rotation) changes. This can be seamlessly combined with Dagger2's `AndroidInjector`.
+
+The demo application shows how to use the library and includes examples on how to use unit and instrumentation tests with it.
 
 ## Quickstart
-1. Add the library
+Add the library:
+```groovy
+implementation 'eu.darken.ommvplib:library:0.2.0'
 ```
-implementation 'eu.darken.ommvplib:library:0.1.0'
+### Without Dagger
+The `Presenter` and the `View` that our `Activity` will implement.
+```java
+@MainComponent.Scope
+public class MainPresenter extends Presenter<MainPresenter.View, MainComponent> {
+
+    @Inject
+    MainPresenter() {
+    }
+
+    @Override
+    public void onBindChange(@Nullable View view) {
+        super.onBindChange(view);
+        onView(v -> doSomething());
+    }
+
+    public interface View extends Presenter.View {
+        void doSomething();
+    }
+}
+
 ```
-2. There is no Quickstart. Look at the example app. It's not as complicated as it looks. 
+
+The process of `attach`ing initialises the `Loader`, checks for any existing `Presenter` and creates one if necessary.
+```java
+public class MainActivity extends AppCompatActivity implements MainPresenter.View {
+
+    MainPresenter presenter;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        OMMVPLib.<MainPresenter.View, MainPresenter>builder()
+                .statePublisher(statePublisher)
+                .presenterCallback(new LoaderFactory.Callback<MainPresenter.View, MainPresenter>() {
+                    @Override
+                    public void onPresenterReady(MainPresenter presenter) {
+                        MainActivity.this.presenter = presenter;
+                    }
+
+                    @Override
+                    public void onPresenterDestroyed() {
+
+                    }
+                })
+                .presenterSource(new PresenterSource<MainPresenter>() {
+                    @Override
+                    public MainPresenter create() {
+                        return new MainPresenter();
+                    }
+                })
+                .attach(this);
+        setContentView(R.layout.activity_main);
+    }
+}
+```
+
+### With Dagger
+The component for our injection.
+```java
+@MainComponent.Scope
+@Subcomponent(modules = {AndroidSupportInjectionModule.class})
+public interface MainComponent extends ActivityComponent<MainActivity>, PresenterComponent<MainPresenter.View, MainPresenter> {
+    
+    @Subcomponent.Builder
+    abstract class Builder extends ActivityComponent.Builder<MainActivity, MainComponent> {}
+    
+    @javax.inject.Scope
+    @Retention(RetentionPolicy.RUNTIME)
+    @interface Scope {}
+}
+```
+
+The `Presenter` that will be injected, including the `View` that our `Activity` will implement.
+
+```java
+@MainComponent.Scope
+public class MainPresenter extends ComponentPresenter<MainPresenter.View, MainComponent> {
+
+    @Inject
+    MainPresenter() {
+    }
+
+    @Override
+    public void onBindChange(@Nullable View view) {
+        super.onBindChange(view);
+        onView(v -> doSomething());
+    }
+
+    public interface View extends Presenter.View {
+        void doSomething();
+    }
+}
+
+```
+
+Now we just need to attach the presenter.
+
+```java
+public class MainActivity extends AppCompatActivity implements MainPresenter.View {
+    
+    @Inject MainPresenter presenter;
+    
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        OMMVPLib.<MainPresenter.View, MainPresenter>builder()
+                .statePublisher(statePublisher)
+                .presenterCallback(new PresenterInjectionCallback<>(this))
+                .presenterSource(new InjectedPresenter<>(this))
+                .attach(this);
+        setContentView(R.layout.activity_main);
+    }
+}
+```
+
 
 ## Acknowledgements
-This library combines and refactored several other concepts. It was based on:
+This library combines multiple concepts: 
 
 * [tomorrow-mvp](https://github.com/michal-luszczuk/tomorrow-mvp) by Michał Łuszczuk
 * [toegether-mvp](https://github.com/laenger/together-mvp) by Christian Langer
